refactor(yield): move the benchmark freshness classifier into shared
The API read path now applies the same benchmark freshness rule as the
cron write path, which crossed the worker api/cron import boundary. The
classifier, the per-key observation bounds, and the record-age helper
move to shared/lib/yield-benchmark-freshness.ts; benchmarks.ts re-exports
them so cron callers are unchanged. Also clears two unused imports and a
dynamic-RegExp lint warning in tests.
