+++ worker/src/cron/__tests__/sync-yield-data-publication-cache.test.ts
@@ -8,7 +8,6 @@ import {
   resetSyncYieldDataTest,
   cleanupSyncYieldDataTest,
   fixtureSyncYieldData,
-  fixtureMockD1,
   fixtureBatchExecute,
   fixtureGetCache,
   fixtureWriteFreshnessSentinel,
